Convert populateDatabase script to TypeScript.
